git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Trunk@74648 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ pff-interface-bajaj/src/main/java/com/pennanttech/bajaj/process/datamart/IVRFlexiDataMart.java
@@ -37,11 +37,11 @@ public class IVRFlexiDataMart extends DatabaseDataEngine implements Runnable {
 
 		parmMap = new MapSqlParameterSource();
 
-		jdbcTemplate.query(sql.toString(), parmMap, new ResultSetExtractor<Integer>() {
+		jdbcTemplate.query(sql.toString(), parmMap, new ResultSetExtractor<Long>() {
 			MapSqlParameterSource map = null;
 
 			@Override
-			public Integer extractData(ResultSet rs) throws SQLException, DataAccessException {
+			public Long extractData(ResultSet rs) throws SQLException, DataAccessException {
 				String[] filterFields = new String[1];
 				filterFields[0] = "AGREEMENTNO";
 				while (rs.next()) {
